Tools/nnengine/requirements.txt
easydict==1.13
numpy==1.24.4
pillow==10.3.0
